requirements.txt
apiclient==1.0.4
asn1crypto==1.5.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cramjam==2.5.0
cryptography==36.0.2
dnspython==1.16.0
fastparquet==0.8.3
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.9
fsspec==2022.8.2
google==3.0.0
google-api-core==2.10.0
google-api-python-client==2.58.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.23.2
oauthlib==3.2.0
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
Pillow==8.1.2
protobuf==4.21.5
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pymongo==3.9.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==2.7.12
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
